requirements.txt
pandas==1.1.4
pyarrow==2.0.0
websocket_client==0.57.0
